+++ app/src/main/java/com/rossyn/blocktiles/game2048/services/Music.java
@@ -4,7 +4,6 @@ import android.content.Intent;
 import android.media.MediaPlayer;
 import android.os.Binder;
 import android.os.IBinder;
-import android.util.Log;
 
 import com.rossyn.blocktiles.game2048.R;